mhm nice. this took 5 hours.

+++ src/MiNET/MiNET/Net/BedrockMessageHandlerBase.cs
@@ -178,7 +178,7 @@ namespace MiNET.Net
 									//if (Log.IsDebugEnabled)
 									//	Log.Debug($"0x{internalBuffer[0]:x2}\n{Packet.HexDump(internalBuffer)}");
 
-									messages.Add(PacketFactory.Create((byte) id, internalBuffer, "mcpe") ??
+									messages.Add(PacketFactory.Create((short) id, internalBuffer, "mcpe") ??
 												new UnknownPacket((byte) id, internalBuffer));
 								}
 								catch (Exception e)
